refactored motions
