+++ src/utils/motionConstants.json
@@ -1,0 +1,25 @@
+{
+    "roll":{
+        "stiffness": 50,
+        "damping": 4,
+        "maxAngle": 30,
+        "axis": "z"
+    },
+    "pitch":{
+        "stiffness": 50,
+        "damping": 4,
+        "maxAngle": 15,
+        "axis": "x"
+    },
+    "yaw":{
+        "rateIncrement": 2,
+        "maxRate": 5,
+        "decayFactor": 0.9,
+        "axis": "y"
+    },
+    "throttle":{
+        "acceleration": 0.5,
+        "maxSpeed": 2,
+        "decayFactor": 0.85
+    }
+}
